removing duplicated class and inserted input check in hamiltonian tomography duration
